Refactored for User based access to resources

+++ src/main/java/com/demo/journalapp/service/JournalEntryService.java
@@ -4,6 +4,8 @@ import com.demo.journalapp.entity.JournalEntry;
 import com.demo.journalapp.entity.User;
 import com.demo.journalapp.repository.JournalEntryRepo;
 import org.bson.types.ObjectId;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -43,21 +45,29 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id).orElse(null);
     }
 
+    @Transactional
     public boolean deleteJournalEntryById(ObjectId id, String userName) {
-        User user = userService.getUserByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id)); //remove a journal entry if the id of the journal entry matches with any id present in the user journal entries list
-        userService.saveUser(user);
-        journalEntryRepo.deleteById(id);
+        try{
+            User user = userService.getUserByUserName(userName);
+            boolean removed = user.getJournalEntries().removeIf(x->x.getId().equals(id)); //remove a journal entry if the id of the journal entry matches with any id present in the user journal entries list
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepo.deleteById(id);
+            }
+        }catch (Exception e){
+            throw new RuntimeException("An error occurred while deleting journal entry" +e.getMessage());
+        }
         return true;
     }
 
-    public JournalEntry updateJournalEntry(ObjectId id, JournalEntry journalEntry, String userName) {
+    public ResponseEntity<JournalEntry> updateJournalEntry(ObjectId id, JournalEntry journalEntry, String userName) {
         JournalEntry oldJournalEntry = journalEntryRepo.findById(id).orElse(null);
         if(oldJournalEntry != null) {
             oldJournalEntry.setTitle(!journalEntry.getTitle().isEmpty() ? journalEntry.getTitle() : oldJournalEntry.getTitle());
             oldJournalEntry.setContent(!journalEntry.getContent().isEmpty() ? journalEntry.getContent() : oldJournalEntry.getContent());
+            return new ResponseEntity<>(journalEntryRepo.save(oldJournalEntry), HttpStatus.ACCEPTED);
         }
-        return journalEntryRepo.save(oldJournalEntry);
+        return ResponseEntity.notFound().build();
     }
 
 }
